reformatted

+++ scripts/locations.py
@@ -1,15 +1,18 @@
-import pandas as pd
-import json
-import os
-import sys
 import argparse
-import multiprocessing
-import sqlite3
+import json
 import logging
+import multiprocessing
+import os
+import sqlite3
+import sys
+
+import pandas as pd
+
 
 def check_path(filename):
-    path = "/".join( filename.split("/")[0:-1] )
-    os.system("mkdir -p " + path)   
+    path = "/".join(filename.split("/")[0:-1])
+    os.system("mkdir -p " + path)
+
 
 class Logger():
 
@@ -25,11 +28,12 @@ class Logger():
     def log(self, level, msg):
         self.root.log(level, msg)
 
-
     def info(self, msg):
         self.root.info(msg)
 
-logger = Logger() 
+
+logger = Logger()
+
 
 def parse_delimiter(delimiter):
     if delimiter == "blanks":
@@ -38,6 +42,7 @@ def parse_delimiter(delimiter):
         return "\t"
     else:
         return ","
+
 
 class Metadata():
 
@@ -61,9 +66,9 @@ class Metadata():
         delimiter = parse_delimiter(delim)
         with open(filename, "r") as filelist:
             for line in filelist:
-                csvPath = line.replace("\n","")
+                csvPath = line.replace("\n", "")
                 print("Reading from", csvPath)
-                frames.append( pd.read_csv(csvPath, delimiter, dtype=dtype, keep_default_na=False) )
+                frames.append(pd.read_csv(csvPath, delimiter, dtype=dtype, keep_default_na=False))
         self.data = pd.concat(frames)
         print("Multiple CSV files loaded")
 
@@ -75,17 +80,19 @@ class Metadata():
         else:
             self.data = self.data.loc[filteringRule(self.data), :]
 
+
 ## Generator of plates. Reads metadata and yields plates
 def read_plates(metaFile):
     metadata = Metadata(metaFile)
     plates = metadata.data["Metadata_Plate"].unique()
     logger.info("Total plates: {}".format(len(plates)))
-    for i in range(len(plates)):  #  & (df.Metadata_Well == "a01")
+    for i in range(len(plates)):  # & (df.Metadata_Well == "a01")
         plate = metadata.filterRecords(lambda df: (df.Metadata_Plate == plates[i]), copy=True)
         yield plate
     return
 
-def print_progress (iteration, total, prefix="Progress", suffix="Complete", decimals=1, barLength=100):
+
+def print_progress(iteration, total, prefix="Progress", suffix="Complete", decimals=1, barLength=100):
     """
     Call in a loop to create terminal progress bar
     @params:
@@ -95,48 +102,61 @@ def print_progress (iteration, total, prefix="Progress", suffix="Complete", deci
         suffix      - Optional  : suffix string (Str)
         decimals    - Optional  : positive number of decimals in percent complete (Int)
         barLength   - Optional  : character length of bar (Int)
-    """ 
-    if all(t >= 0 for t in [iteration,total,barLength]) and iteration <= total:
-        formatStr       = "{0:." + str(decimals) + "f}"
-        percents        = formatStr.format(100 * (iteration / float(total)))
-        filledLength    = int(round(barLength * iteration / float(total)))
-        bar             = "#" * filledLength + "-" * (barLength - filledLength)
+    """
+    if all(t >= 0 for t in [iteration, total, barLength]) and iteration <= total:
+        formatStr = "{0:." + str(decimals) + "f}"
+        percents = formatStr.format(100 * (iteration / float(total)))
+        filledLength = int(round(barLength * iteration / float(total)))
+        bar = "#" * filledLength + "-" * (barLength - filledLength)
         sys.stdout.write("\r%s |%s| %s%s %s" % (prefix, bar, percents, "%", suffix)),
         sys.stdout.flush()
         if iteration == total:
             sys.stdout.write("\n")
             sys.stdout.flush()
-    elif sum([iteration<0,total<0,barLength<0]) > 1:
+    elif sum([iteration < 0, total < 0, barLength < 0]) > 1:
         sys.stdout.write("\rError: print_progress() function received multiple negative values.")
         sys.stdout.flush()
     elif iteration < 0:
-        sys.stdout.write("\rError: print_progress() function received a negative "iteration" value.")
+        sys.stdout.write("\rError: print_progress() function received a negative "
+        iteration
+        " value.")
         sys.stdout.flush()
     elif total < 0:
-        sys.stdout.write("\rError: print_progress() function received a negative "total" value.")
+        sys.stdout.write("\rError: print_progress() function received a negative "
+        total
+        " value.")
         sys.stdout.flush()
     elif barLength < 0:
-        sys.stdout.write("\rError: print_progress() function received a negative "barLength" value.")
+        sys.stdout.write("\rError: print_progress() function received a negative "
+        barLength
+        " value.")
         sys.stdout.flush()
     elif iteration > total:
-        sys.stdout.write("\rError: print_progress() function received an "iteration" value greater than the "total" value.")
+        sys.stdout.write("\rError: print_progress() function received an "
+        iteration
+        " value greater than the "
+        total
+        " value.")
         sys.stdout.flush()
+
 
 def write_locations(field, query_template, plate_name, row, conn, config):
     # Read cells file for each image
-    query = query_template.replace("@@@",field).format(
-            plate_name,
-            row["Metadata_Well"],
-            row["Metadata_Site"]
+    query = query_template.replace("@@@", field).format(
+        plate_name,
+        row["Metadata_Well"],
+        row["Metadata_Site"]
     )
     locations = pd.read_sql_query(query, conn)
 
     # Keep center coordinates only, remove NaNs, and transform to integers
     locations = locations.dropna(axis=0, how="any")
-    locations[field+"_Location_Center_X"] = locations[field+"_Location_Center_X"]*config["compression"]["scaling_factor"]
-    locations[field+"_Location_Center_Y"] = locations[field+"_Location_Center_Y"]*config["compression"]["scaling_factor"]
-    locations[field+"_Location_Center_X"] = locations[field+"_Location_Center_X"].astype(int)
-    locations[field+"_Location_Center_Y"] = locations[field+"_Location_Center_Y"].astype(int)
+    locations[field + "_Location_Center_X"] = locations[field + "_Location_Center_X"] * config["compression"][
+        "scaling_factor"]
+    locations[field + "_Location_Center_Y"] = locations[field + "_Location_Center_Y"] * config["compression"][
+        "scaling_factor"]
+    locations[field + "_Location_Center_X"] = locations[field + "_Location_Center_X"].astype(int)
+    locations[field + "_Location_Center_Y"] = locations[field + "_Location_Center_Y"].astype(int)
 
     # Save the resulting dataset frame in the output directory
     loc_file = "{}/{}/locations/{}-{}-{}.csv".format(
@@ -152,21 +172,21 @@ def write_locations(field, query_template, plate_name, row, conn, config):
 
 def create_cell_indices(args):
     plate, config = args
- 
+
     # Open database
     plate_name = plate.data.iloc[0]["Metadata_Plate"]
     database_file = "{}/{}/{}.sqlite".format(config["original_images"]["backend"], plate_name, plate_name)
     conn = sqlite3.connect(database_file)
 
     # Define query template: @@@ is either Cells or Nuclei
-    query_template = "SELECT @@@_Location_Center_X, @@@_Location_Center_Y " +\
-                     " FROM @@@ INNER JOIN Image " +\
-                     "    ON Image.ImageNumber = @@@.ImageNumber " +\
-                     "    AND Image.TableNumber = @@@.TableNumber " +\
-                     " WHERE Image.Image_Metadata_Plate = '{}' " +\
-                     "    AND Image.Image_Metadata_Well = '{}' " +\
-                     "    AND Image.Image_Metadata_Site = '{}' " +\
-                     "    AND @@@_Location_Center_X NOT LIKE 'NaN' " +\
+    query_template = "SELECT @@@_Location_Center_X, @@@_Location_Center_Y " + \
+                     " FROM @@@ INNER JOIN Image " + \
+                     "    ON Image.ImageNumber = @@@.ImageNumber " + \
+                     "    AND Image.TableNumber = @@@.TableNumber " + \
+                     " WHERE Image.Image_Metadata_Plate = '{}' " + \
+                     "    AND Image.Image_Metadata_Well = '{}' " + \
+                     "    AND Image.Image_Metadata_Site = '{}' " + \
+                     "    AND @@@_Location_Center_X NOT LIKE 'NaN' " + \
                      "    AND @@@_Location_Center_Y NOT LIKE 'NaN' "
 
     # Extract cells and nuclei locations for each image
@@ -178,19 +198,21 @@ def create_cell_indices(args):
         iteration += 1
     print("")
 
+
 class Parallel():
 
     def __init__(self, fixed_args, numProcs=None):
         self.fixed_args = fixed_args
-        cpus =  multiprocessing.cpu_count()
+        cpus = multiprocessing.cpu_count()
         if numProcs is None or numProcs > cpus or numProcs < 1:
             numProcs = cpus
         self.pool = multiprocessing.Pool(numProcs)
 
     def compute(self, operation, data):
-        iterable = [ [d, self.fixed_args] for d in data ]
+        iterable = [[d, self.fixed_args] for d in data]
         self.pool.map(operation, iterable)
         return
+
 
 parser = argparse.ArgumentParser(description="Find cell locations")
 parser.add_argument("config", help="The path to the configuration file")
@@ -201,5 +223,5 @@ cores = int(options.core)
 with open(options.config, "r") as f:
     config = json.load(f)
 process = Parallel(config, numProcs=cores)
-metadata = read_plates(config["metadata"]["path"]+config["metadata"]["filename"])
+metadata = read_plates(config["metadata"]["path"] + config["metadata"]["filename"])
 process.compute(create_cell_indices, metadata)
